app/requirements.txt
Django==3.2.5
gunicorn==20.0.4
psycopg2-binary==2.8.5
django-rest-swagger==2.2.0
djangorestframework==3.11.2
PyYAML
six
djangorestframework-api-key
drf-yasg==1.17.1
django-allauth==0.42.0
django-rest-registration==0.6.3
django-rest-auth==0.9.5
packaging